Fix the protocol_splitter receive buffer getting full case
Always make room in the receive buffer for reading out data. Otherwise the
client may end up in busyloop, when there is no valid data in the buffer to read.
The poll may still return 1, in case there is new data in the actual input
device (socket or uart buffer).

Signed-off-by: Jukka Laitinen <jukkax@ssrc.tii.ae>