boards/xx/toc: Move the ToC to the beginning of the firmware image
This allows signinig the protected mode binary, by encapsulating the
firmware between a header (the ToC) and the signature, instead of placing
the ToC inside of the firmware itself.

The problem was that the ToC went to the kernel binary, while the
signature is appended to the user binary. The ToC needs the location of
the signature which cannot be known at the time when the kernel is linked.

For ToC integrity protection, the ToC itself is also signed and verified.

One clear optimization would be to remove the need for the signed toc
truncation part, currently a fixed size (512B) ToC area is used. This
can be optimized to use a 'perfect size fit' ToC, but requires a bit
of linker magic (use a dummy ToC + signature to occupy memory from
the firmware build), but this can be done later.