boards/xx/toc: Put ToC to the beginning of the firmware image
This allows signinig the protected mode binary, by encapsulating the
firmware between a header (the ToC) and the signature, instead of placing
the ToC inside of the firmware itself.

The problem was that the ToC went to the kernel binary, while the
signature is appended to the user binary. The ToC needs the location of
the signature which cannot be known at the time when the kernel is linked.

This fix reserves an are for the ToC from the beginning of the program
memory:
- The first element is the ToC
- After the ToC, the _unsigned_ px4 binary is injected

Finally, a binary file containing the ToC + px4 firmware is created, and
this is signed, which is the final firmware image.