icicle: Put ToC to the beginning of the firmware image
This allows signinig the protected mode binary, as the requirement for the
linker defined symbol "_boot_signature" is removed. The problem is that
the ToC goes to the kernel binary, while the signature is appended to
the user binary, but the ToC needs the location of the signature which
cannot be known when the kernel is linked.

This fix reserves 256 bytes from the beginning of the program memory
for the ToC (which is more like a firmware header now):
- The first element is a trampoline to the application. The bootloader
  assumes that the application head is the beginning of the program
  memory.
- The second element is the boot delay signature (do we even use this ?)
- The third element is the ToC
- After the ToC, the _unsigned_ px4 binary is injected

Finally, a binary file containing the ToC + px4 firmware is created, and
this is signed, which is the final firmware image.

It is still unclear whether this approach will work for ARM targets, which
expect the reset vector at a certain location (usually, beginning of FLASH).