boards/xx/toc: Move the ToC to the beginning of the firmware image
This allows signinig the protected mode binary, by encapsulating the
firmware between a header (the ToC) and the signature, instead of placing
the ToC inside of the firmware itself.

The problem was that the ToC went to the kernel binary, while the
signature is appended to the user binary. The ToC needs the location of
the signature which cannot be known at the time when the kernel is linked.

For ToC integrity protection, the ToC itself is also signed and verified.