mpfs/bootloader: Fix bootloader softlock if SD card is missing
If the SD card is not mounted, the bootloader will softlock in
flash_func_erase_sector() where loading_status is tested, and if
the value is not FAIL/DONE the bootloader_main process will sleep
forever.

This happens because the task that was supposed to wake it up (loader_main)
has already exited because it had no work to be done due to the SD card
missing.