Fix CONFIG_BUILD_PROTECTED after NuttX update
The syscall lookup table contains poisoned xxenv() procedures and causes
a build error. The file is included via a massive header that includes
almost everything in /arch/risc-v/ etc (riscv_internal.h). This was
already fixed once, and, tbh, suppressing the poison error is the only
way to get this to work reliably.