board_ioctl: Remove dependency to lib/builtin.h
The only thing needed from there was the definition of struct builtin_s.
The builtin functions are not meant to be used this way anyhow, it is
supposed to be a way to inject user-space code into the kernel (the
builtin list is _copied_ into kernel memory). The board_ioctl module is
is kernel-space only.

Also, remove the usage of the "FAR" keyword, we use processors from this
millennium that do not use memory segmentation into 64K chunks...

The real reason for this change is that CONFIG_BUILD_KERNEL does not
allow setting CONFIG_BUILTIN (makes sense, injecting user code into the
kernel is not feasible _at all_ with address environments).