nuttx/config_build_kernel: Add support for building NuttX in kernel mode
This is phase 0.1 for enabling PX4 in NuttX kernel mode, this builds
the simplest possible configuration i.e. NuttX + nsh + px4 userspace init,
but no userspace processes yet.

One result is a bootable ROMFS which is placed into the kernel binary
and mounted into /sbin by the board initialization logic. The kernel
knows the mount point and starts the user space from /sbin/init, which
is the nsh shell.

A second result are the NuttX apps, placed and installed into /bin

The next step is to start building the modules into processes