nuttx/config_build_kernel: Build the modules into processes
This commit adds support to build px4 modules into individual processes.
This is done by using a common main.cpp file, which calls the user space
init function and then the module entry point.

The entry point is provided by a command line CDEF -Dentry=module_main
The processes are installed into /bin and this is mounted as from a ROMFS

The dependencies are still a huge mess, and will be fixed later...