px4_init: Fix linker defined symbols after NuttX update
For some reason, the type of the linker defined symbols was changed.
The reasoning was "they typically represent pointers" which is absolutely
not true at all. They typically define memory locations, but they can
be whatever. It is just a way to provide symbols from the linker to
the executable, which is done via relocation symbols.