nuttx/config_build_kernel: Do not clear environment in kernel
The environment is cleared for px4 processes, but the initial process
needs to have the environment intact. Without this patch the environment
is wiped for nsh (the init process) via:

px4_platform_init ->
px4::WorkQueueManagerStart(); ->
px4_task_spawn_cmd() ->
clearenv()

nsh _NEEDS_ the environment to be intact in kernel mode.