nuttx/config_build_kernel: Use symbolic links for starting kernel modules
The proxy process "launch_kmod" is used to start the kernel side modules,
to keep backwards compatibility with nsh builtin apps, use symbolic links
to achive the same functionality.

rc.sysinit is used for this, it is perfect for setting up low level
infrastructure needed by the init process, as it is loaded _before_ the
rcS / init.d script.