bootloader/microchip/mpfs/main.c: Fix POR reset
The debugger reset reason is giving random values at high temperaturs in power-on-reset.

Change the bootloader logic to first check if the device has been cold-booted and only after
that check the debugger reset reason. Use the fabric reset reason to detect POR, since that is
only set at power-on. Other bits don't give the correct status, the specs are buggy.

Signed-off-by: Jukka Laitinen <jukkax@ssrc.tii.ae>