hrt_ioctl: Do not double post same entry to inflight queue
Make sure the same entry is not put into inflight queue twice.

Why? Because the list is a singly linked list, and posting the same
entry twice would create a queue loop which soft locks the entire system.

When does this happen ? If the client thread cannot read the event out
in time, a new event can arrive. The same handle is used for HRT_CALL_EVERY
and in this case double booking can happen.