drivers/ads1115: Fix consecutive warning "ADS1115 not ready!"
ADS1115 is first commanded to start sampling and after 1/4 of the sample
period the single channel sample is read out.

However, when using ScheduleOnInterval this fails if the sample reading
is blocked for long enough.

The resulting _average_ sample rate is still correct (80Hz) but the time
between triggering the new sample on the ADC and reading it out can be
anything. If the call to ADS1115::RunImpl() happens too soon after starting
the new sample, reading it will obviously fail.

Fix this by scheduling the sample reading _after_ the sampling has been
triggered on the ADC. This ensures the sample is ready the next time when
ADS1115::RunImpl() is executed.