mavlink/mavlink_shell: Move mavlink shell to its own process in NuttX build
Move the mavlink shell to its own process and launch it via px4_exec.

The task is very simple and does not need anything from the MavlinkShell
class, it is defined as a static function anyway in order to be passed
as pointer-to-function to px4_task_spawn.

The shell application is now compiled as its own module, and added to
the builtin list or list of module binaries, and can / is started via
px4_exec instead.