Add more interfaces for PX4Crypto
Add renew_nonce and and decrypt_data interfaces, and also ioctl's for memory
protected builds. Stub these functions in sw_crypto driver. the functions are
currently not used but a stub is needed in order to build properly.

Add implementation for signature_check in PX4Crypto class, there is already an
implementation in the backend, but the function is missing from the PX4Crypto
interface.

Signed-off-by: Jukka Laitinen <jukkax@ssrc.tii.ae>