SubscriptionCallback: Add lock protecting callback during its execution
There is a potential problem with sub->call() and sub->[un-]register(),
when the callback is executing, if someone drops the callback registration,
it will result in a resource leak and a crash.

This does not happen in flat mode, as the uORB::DeviceNode::lock protects
both, but when a callback thread is used, there is a race condition which
is fixed by this patch.