Component: Introduced failure_detector_hitl module, integrating three classes for handling failure injection messages(FailureDetectorHITL), simulating sensors that have ceased to update (FakeStuckSensor), and a component managing a collection of these simulated sensors (FakeSensors).
This module is a combination of three classes: FailureDetectorHITL for detecting failure injection messages,
FakeStuckSensor for simulating sensors whose readings have become stuck, and FakeSensors, a component that aggregates a collection of such simulated sensors.

Additionally, the handling of the sensors module has been extended to support disabling sensors during flight.

Designed exclusively for operation in HITL (Hardware-In-The-Loop) mode, this module does not interfere with normal operational mode.

Reported-by: Ilia Loginov ilia.loginov@tii.ae

On branch FEATURE_FAILURE_INJECTION_HITL
Changes to be committed:
	modified:   src/modules/sensors/CMakeLists.txt
	new file:   src/modules/sensors/failure_detector_HITL/CMakeLists.txt
	new file:   src/modules/sensors/failure_detector_HITL/failure_detector_HITL.cpp
	new file:   src/modules/sensors/failure_detector_HITL/failure_detector_HITL.hpp
	modified:   src/modules/sensors/sensors.cpp
	modified:   src/modules/sensors/sensors.hpp