mpfs/px4io_serial.cpp: Remove unnecessary critical section
Masking all system interrupts in this case is not necessarily a good idea
as the csection can be relatively long due to accessing a low speed
peripheral domain, potentially writing a lot of data via the slow
BUS.

I presume the original code tries to prevent any interrupt from executing
during the FIFO fill, but this will also starve all other interrupt sources
for the duration of the FIFO fill, so it should not be considered a valid
way to ensure atomicy of the px4io message.

The best solution for this issue would be to use DMA for the transfer
(don't know if DMA is available).

Mitigate the aforementioned issue by enabling the px4io RX interrupt only
after the TX FIFO has been fully written.