platforms/nuttx: Remove unnecessary sched_lock() from task spawn logic
Starting a new task does not require that the scheduler is locked, this is
done in nsh only because nsh uses waitpid() to wait for the task to exit
and record its exit status.

The exit status does not interest us whatsoever -> remove the locking.

Also, we don't need to explicitly change a tasks priority when it is
created by px4_task_spawn_cmd, as task_create() already sets the priority
via nxthread_setup_scheduler().