ADIS16470: Fix reset and self-test sequence
Fix issues related to unspecified delays in the ADIS sensor:

1. The reset occurs only after some delay from writing the SW reset, and the 193ms
mentioned in the datasheet occurs only after that. When the reset sequence really
starts can be determined with logic analyzer by monitoring DR signal and SPI writes.
This delay is measured to be between 10-20ms, so add 20ms to SW reset recovery time.

2. The self test sequence starts only after some delay from writing the command, and the
14ms mentioned in the datasheet starts only after this delay. Again, the behaviour can be
studied with logic analyzer by monitoring the DR signal and the SPI writes - when the self
test is completed, the IMU starts measuring again and the DR starts toggling. This
unspecified delay is measured to be somewhere between 5-10ms, so add 10ms to self test time.

3. After writing the register set in Configure(), there is an unspecified delay before the
register can be read back. This was bisected to be around 50us, so add 100us delay after
setting the registers before trying to read them back.

In addition:
1. Make sure that the IMU has woken up after reset by reading the PROD_ID register
always after the reset, also before starting the self test sequence on the first boot.
2. If self test fails, reset the IMU as a false result may occur due to moving the device
during self test (according to the datasheet).

Signed-off-by: Jukka Laitinen <jukkax@ssrc.tii.ae>