nuttx/CMakeLists.txt: Fix dependency for boot/init (again)
The dependency should be to nuttx_app_bins target, since nsh is deleted
from the /bin folder (to not allow users to execute it). Otherwise the
nsh target will always be re-build (as it does not exist). Using
nuttx_app_bins as the dependency fixes it, and when the content of nsh
is changed it also re-triggers all relevant targets.