modules/sensors: Fix a race condition in VehicleIMU::UpdateAccel
First checking if the accelerometer is updated in VehicleIMU::Run and later updating
it in UpdateAccel is a race; as there is deviation between the _accel_interval_us and the
real update interval, it is possible to unnecessarily start to skip every second accelerometer
sample if the update happens right afer checking if it was updated.

Remove separate check of "_sensor_accel_sub.updated" and just update it when the other
conditions for update are met.

Signed-off-by: Jukka Laitinen <jukka.laitinen@tii.ae>