platforms/nuttx: Add maximum page size for loadable elf module linking
Set "-z max-page-size=0x1000" for all loadable applications in CONFIG_BUILD_KERNEL.

This is the only page size supported by nuttx, and saves 60KB of space for every
application elf file for aarch64 compilation, which defaults to 64KB page size.

With larger page size, the .text segment gets always aligned to 64K boundary, and
the beginning of the elf file is padded with 0.

Signed-off-by: Jukka Laitinen <jukka.laitinen@tii.ae>