platforms/imx9/hrt: In hrt_call_invoke, keep g_hrt_lock locked while processing the call
When calling the callback fucnction, the g_hrt_lock was free during the call, and after the
call it was locked again before re-entering a peridic callback.

This has the issue that on an SMP system the call can be re-scheduled or cancelled
on another CPU while calling the callback on the other, causing the call to be
invalid/corrupt at the point when the periodic call is re-scheduled.

Fix this by first re-scheduling the periodic call before calling the callback.

Signed-off-by: Jukka Laitinen <jukka.laitinen@tii.ae>